feat: implement `Zicsr` extension

+++ src/extensions/rv32i.rs
@@ -402,7 +402,7 @@ impl crate::Instruction for Instruction {
         exec(self, cpu)
     }
 
-    fn len(&self) -> usize {
+    fn len(&self) -> u32 {
         4
     }
 }
